docs(README): mention how to optimize the size/startup time
We mention several options that can be used and `min-sized-rust` that documents them. It can help users know how to reduce the size of their builds and decrease the startup time.

In the future we might enable by default some of these options, but in the mean time we should at least document it.